Fixes for jaxtyping to work

+++ grad_dft/molecule.py
@@ -27,7 +27,6 @@ from flax import linen as nn
 import jax
 
 from jaxtyping import Array, PyTree, Scalar, Float, Int
-from typeguard import typechecked
 
 
 @struct.dataclass
@@ -279,7 +278,6 @@ class Molecule:
 
 #######################################################################
 
-#@typechecked
 def orbital_grad(
         mo_coeff: Float[Array, "spin orbitals orbitals"],
         mo_occ: Float[Array, "spin orbitals"],
@@ -319,7 +317,6 @@ def orbital_grad(
 
 ##########################################################
 
-#@typechecked
 @partial(jax.jit, static_argnames="precision")
 def density(rdm1: Float[Array, "spin orbitals orbitals"], 
             ao: Float[Array, "grid orbitals"], 
@@ -344,7 +341,6 @@ def density(rdm1: Float[Array, "spin orbitals orbitals"],
 
     return jnp.einsum("...ab,ra,rb->r...", rdm1, ao, ao, precision=precision)
 
-#@typechecked
 @partial(jax.jit, static_argnames="precision")
 def grad_density(
     rdm1: Float[Array, "spin orbitals orbitals"], 
@@ -374,7 +370,6 @@ def grad_density(
 
     return 2 * jnp.einsum("...ab,ra,rbj->r...j", rdm1, ao, grad_ao, precision=precision)
 
-#@typechecked
 @partial(jax.jit, static_argnames="precision")
 def lapl_density(
     rdm1: Float[Array, "spin orbitals orbitals"], 
@@ -407,7 +402,6 @@ def lapl_density(
         "...ab,raj,rbj->r...", rdm1, grad_ao, grad_ao, precision=precision
     ) + 2 * jnp.einsum("...ab,ra,rbi->r...", rdm1, ao, grad_2_ao, precision=precision)
 
-#@typechecked
 @partial(jax.jit, static_argnames="precision")
 def kinetic_density(
     rdm1: Float[Array, "spin orbitals orbitals"], 
@@ -434,7 +428,6 @@ def kinetic_density(
 
     return 0.5 * jnp.einsum("...ab,raj,rbj->r...", rdm1, grad_ao, grad_ao, precision=precision)
 
-#@typechecked
 @partial(jax.jit, static_argnames=["precision"])
 def HF_energy_density(
     rdm1: Float[Array, "spin orbitals orbitals"],
@@ -472,7 +465,6 @@ def HF_energy_density(
     )
     return vmap(_hf_energy, in_axes=(0, None, 0), out_axes=2)(chi, rdm1, ao)
 
-#@typechecked
 def HF_density_grad_2_Fock(
     grid: Grid,
     functional: nn.Module,
@@ -543,7 +535,6 @@ def HF_density_grad_2_Fock(
 
     return (jax.jit(chunked_jvp)(chi.transpose(3, 0, 1, 2), gr, ao)).transpose(1, 2, 3, 0)
 
-#@typechecked
 def HF_coefficient_input_grad_2_Fock(
     grid: Grid,
     functional: nn.Module,
@@ -620,7 +611,6 @@ def abs_clip(arr, threshold):
 
 ######################################################################
 
-#@typechecked
 @partial(jax.jit, static_argnames=["precision"])
 def nonXC(
     rdm1: Float[Array, "spin orbitals orbitals"],
@@ -658,7 +648,6 @@ def nonXC(
 
     return nuclear_repulsion + h1e_energy + coulomb2e_energy
 
-#@typechecked
 @partial(jax.jit)
 def symmetrize_rdm1(rdm1: Float[Array, "spin orbitals orbitals"]) -> Float[Array, "spin orbitals orbitals"]:
     r"""A function that symmetrizes and clips the reduced density matrix.
@@ -676,7 +665,6 @@ def symmetrize_rdm1(rdm1: Float[Array, "spin orbitals orbitals"]) -> Float[Array
     dm = abs_clip(dm, 1e-20)
     return jnp.stack([dm, dm], axis=0) / 2.0
 
-#@typechecked
 @partial(jax.jit, static_argnames=["precision"])
 def coulomb_energy(
     rdm1: Float[Array, "spin orbitals orbitals"],
@@ -701,7 +689,6 @@ def coulomb_energy(
     coulomb2e_energy = jnp.einsum("sji,sij->", rdm1, v_coul, precision=precision) / 2.0
     return coulomb2e_energy
 
-#@typechecked
 @partial(jax.jit, static_argnames=["precision"])
 def one_body_energy(
     rdm1: Float[Array, "spin orbitals orbitals"],
@@ -725,7 +712,6 @@ def one_body_energy(
     h1e_energy = jnp.einsum("sij,ji->", rdm1, h1e, precision=precision)
     return h1e_energy
 
-#@typechecked
 def coulomb_potential(
     rdm1: Float[Array, "spin orbitals orbitals"],
     rep_tensor: Float[Array, "orbitals orbitals orbitals orbitals"],
@@ -752,7 +738,6 @@ def coulomb_potential(
     # The 2 is to compensate for the /2 in the dm definition
     return 2 * jnp.einsum("pqrt,srt->spq", rep_tensor, rdm1, precision=precision)
 
-#@typechecked
 def make_rdm1(
     mo_coeff: Float[Array, "spin orbitals orbitals"],
     mo_occ: Float[Array, "spin orbitals"]
@@ -784,12 +769,11 @@ def make_rdm1(
 
     return jnp.einsum("sij,sj,skj -> sik", mo_coeff, mo_occ, mo_coeff.conj())
 
-##@typechecked #todo Exception Do not use `isinstance(x, jaxtyping.Int)`. If you want to check just the dtype of an array, then use `jaxtyping.Int[jnp.ndarray, "..."]`.
 def get_occ(
     mo_energies: Float[Array, "spin orbitals"],
     nelecs: Int[Array, "spin"],
-    naos: Int,
-) -> Int[Array, "spin orbitals"]:
+    naos: int,
+) -> Float[Array, "spin orbitals"]:
     r""" Computes the occupation of the molecular orbitals.
 
     Parameters
